ai_trading_bot/requirements.txt
numpy>=1.24.0,<2.0.0
requests>=2.31.0,<3.0.0
websocket-client>=1.6.0,<2.0.0
pyyaml>=6.0,<7.0.0
python-dotenv>=1.0.0,<2.0.0

